test(backend): adicionar testes abrangentes para atividades de cards
- Criar testes para autenticação de atividades
- Adicionar testes para criação, atualização, exclusão e movimentação de cards
- Implementar testes de rate limiting para atividades
- Configurar setup compartilhado para testes de atividades
- Melhorar cobertura de testes para operações críticas de cards
